+++ packages/preview-middleware-client/src/adp/init-dialogs.ts
@@ -12,9 +12,11 @@ import FlUtils from 'sap/ui/fl/Utils';
 import type ElementOverlay from 'sap/ui/dt/ElementOverlay';
 
 import ManagedObject from 'sap/ui/base/ManagedObject';
-import { isReuseComponent } from '../cpe/utils';
-import { Ui5VersionInfo } from '../utils/version';
 import { DialogFactory, DialogNames } from './dialog-factory';
+import type { IsReuseComponentApi } from '../cpe/types';
+import { getTextBundle, type TextBundle } from '../i18n';
+import { getReuseComponentChecker } from './utils';
+import type { Ui5VersionInfo } from '../utils/version';
 import { isLowerThanMinimalUi5Version } from '../utils/version';
 
 /**
@@ -22,7 +24,7 @@ import { isLowerThanMinimalUi5Version } from '../utils/version';
  *
  * @param control UI5 control.
  * @param syncViewsIds Runtime Authoring
- * @param ui5VersionInfo UI5 version information
+ * @param isReuseComponent Function to check if the control is a reuse component.
  * @param isCloud Whether the application is running in the cloud
  *
  * @returns boolean whether menu item is enabled or not
@@ -30,14 +32,14 @@ import { isLowerThanMinimalUi5Version } from '../utils/version';
 export function isControllerExtensionEnabledForControl(
     control: ManagedObject,
     syncViewsIds: string[],
-    ui5VersionInfo: Ui5VersionInfo,
+    isReuseComponent: IsReuseComponentApi,
     isCloud: boolean
 ): boolean {
-    const clickedControlId = FlUtils.getViewForControl(control).getId();
-    const isControlInSyncView = syncViewsIds.includes(clickedControlId);
+    const viewId = FlUtils.getViewForControl(control).getId();
+    const isControlInSyncView = syncViewsIds.includes(viewId);
 
     if(isCloud) {
-        const isClickedControlReuseComponent = isReuseComponent(clickedControlId, ui5VersionInfo);
+        const isClickedControlReuseComponent = isReuseComponent(control.getId());
         return !isControlInSyncView && !isClickedControlReuseComponent;
     }
     return !isControlInSyncView;
@@ -48,7 +50,7 @@ export function isControllerExtensionEnabledForControl(
  *
  * @param overlays Control overlays
  * @param syncViewsIds Runtime Authoring
- * @param ui5VersionInfo UI5 version information
+ * @param isReuseComponent Function to check if the control is a reuse component.
  * @param isCloud Whether the application is running in the cloud
  *
  * @returns boolean whether menu item is enabled or not
@@ -56,44 +58,72 @@ export function isControllerExtensionEnabledForControl(
 export const isControllerExtensionEnabled = (
     overlays: ElementOverlay[],
     syncViewsIds: string[],
-    ui5VersionInfo: Ui5VersionInfo,
+    isReuseComponent: IsReuseComponentApi,
     isCloud: boolean
 ): boolean => {
     if (overlays.length === 0 || overlays.length > 1) {
         return false;
     }
-    return isControllerExtensionEnabledForControl(overlays[0].getElement(), syncViewsIds, ui5VersionInfo, isCloud);
+    return isControllerExtensionEnabledForControl(overlays[0].getElement(), syncViewsIds, isReuseComponent, isCloud);
 };
 
 /**
  * Determines whether the fragment command should be enabled based on the provided overlays.
  *
  * @param {ElementOverlay[]} overlays - An array of ElementOverlay objects representing the UI overlays.
- * @param ui5VersionInfo UI5 version information
+ * @param {isReuseComponentApi} isReuseComponent - Function to check if the control is a reuse component.
+ * @param {boolean} isCloud - Whether the application is running in the cloud.
  * @returns {boolean} True if the fragment command is enabled, false otherwise.
  */
-export const isFragmentCommandEnabled = (overlays: ElementOverlay[], ui5VersionInfo: Ui5VersionInfo): boolean => {
+export const isFragmentCommandEnabled = (overlays: ElementOverlay[], isReuseComponent: IsReuseComponentApi, isCloud: boolean): boolean => {
     if (overlays.length === 0 || overlays.length > 1) {
         return false;
     }
 
     const control = overlays[0].getElement();
+    const stableId = hasStableId(overlays[0]);
+    if (isCloud) {
+        return stableId && !isReuseComponent(control.getId());
+    }
 
-    return hasStableId(overlays[0]) && !isReuseComponent(control.getId(), ui5VersionInfo);
+    return stableId;
 };
 
 /**
  * Determines the text that should be displayed for the Add Fragment context menu item.
  *
  * @param {ElementOverlay} overlay - An ElementOverlay object representing the UI overlay.
+ * @param {isReuseComponentApi} isReuseComponentChecker - Function to check if the control is a reuse component.
+ * @param {boolean} isCloud - Whether the application is running in the cloud.
+ * @param {TextBundle} resources - The text bundle.
  * @returns {string} The text of the Add Fragment context menu item.
  */
-export const getAddFragmentItemText = (overlay: ElementOverlay) => {
+export const getAddFragmentItemText = (overlay: ElementOverlay, isReuseComponentChecker: IsReuseComponentApi, isCloud: boolean, resources: TextBundle) => {
+    if (isCloud && isReuseComponentChecker(overlay.getElement().getId())) {
+        return resources.getText('ADP_ADD_FRAGMENT_MENU_ITEM_REUSE_COMPONENT');
+    }
     if (!hasStableId(overlay)) {
-        return 'Add: Fragment (Unavailable due to unstable ID of the control or its parent control)';
+        return resources.getText('ADP_ADD_FRAGMENT_MENU_ITEM_UNSTABLE_ID');
     }
 
-    return 'Add: Fragment';
+    return resources.getText('ADP_ADD_FRAGMENT_MENU_ITEM');
+};
+
+/**
+ * Determines the text that should be displayed for Controller Extension context menu item.
+ *
+ * @param {ElementOverlay} overlay - An ElementOverlay object representing the UI overlay.
+ * @param {isReuseComponentApi} isReuseComponentChecker - Function to check if the control is a reuse component.
+ * @param {boolean} isCloud - Whether the application is running in the cloud.
+ * @param {TextBundle} resources - The text bundle.
+ * @returns {string} The text of the Add Fragment context menu item.
+ */
+export const getExtendControllerItemText = (overlay: ElementOverlay, isReuseComponentChecker: IsReuseComponentApi, isCloud: boolean, resources: TextBundle) => {
+    if (isCloud && isReuseComponentChecker(overlay.getElement().getId())) {
+        return resources.getText('ADP_ADD_CONTROLLER_EXTENSION_MENU_ITEM_REUSE_COMPONENT');
+    }
+
+    return resources.getText('ADP_ADD_CONTROLLER_EXTENSION_MENU_ITEM');
 };
 
 /**
@@ -103,18 +133,20 @@ export const getAddFragmentItemText = (overlay: ElementOverlay) => {
  * @param syncViewsIds Ids of all application sync views
  * @param ui5VersionInfo UI5 version information
  */
-export const initDialogs = async (rta: RuntimeAuthoring, syncViewsIds: string[], ui5VersionInfo: Ui5VersionInfo): Promise<void> => {
+export const initDialogs = async (rta: RuntimeAuthoring, syncViewsIds: string[], ui5VersionInfo: Ui5VersionInfo): Promise<Promise<void>> => {
     const contextMenu = rta.getDefaultPlugins().contextMenu;
     const isCloud = rta.getFlexSettings().isCloud;
-    
+        const resources = await getTextBundle();
+    const isReuseComponentChecker = await getReuseComponentChecker(ui5VersionInfo);
+
     if (isLowerThanMinimalUi5Version(ui5VersionInfo, { major: 1, minor: 134 })) {
         contextMenu.addMenuItem({
             id: 'ADD_FRAGMENT',
-            text: getAddFragmentItemText,
+            text: (overlay: ElementOverlay) => getAddFragmentItemText(overlay, isReuseComponentChecker, isCloud, resources),
             handler: async (overlays: UI5Element[]) =>
                 await DialogFactory.createDialog(overlays[0], rta, DialogNames.ADD_FRAGMENT),
             icon: 'sap-icon://attachment-html',
-            enabled: (overlays: ElementOverlay[]) => isFragmentCommandEnabled(overlays, ui5VersionInfo)
+            enabled: (overlays: ElementOverlay[]) => isFragmentCommandEnabled(overlays, isReuseComponentChecker, isCloud)
         });
     } else {
         const addFragmentService = (await import('open/ux/preview/client/adp/add-fragment')).default;
@@ -124,11 +156,11 @@ export const initDialogs = async (rta: RuntimeAuthoring, syncViewsIds: string[],
     if (isLowerThanMinimalUi5Version(ui5VersionInfo, { major: 1, minor: 135 })) {
         contextMenu.addMenuItem({
             id: 'EXTEND_CONTROLLER',
-            text: 'Extend With Controller',
+            text: (overlay: ElementOverlay) => getExtendControllerItemText(overlay, isReuseComponentChecker, isCloud, resources),
             handler: async (overlays: UI5Element[]) =>
                 await DialogFactory.createDialog(overlays[0], rta, DialogNames.CONTROLLER_EXTENSION),
             icon: 'sap-icon://create-form',
-            enabled: (overlays: ElementOverlay[]) => isControllerExtensionEnabled(overlays, syncViewsIds, ui5VersionInfo, isCloud)
+            enabled: (overlays: ElementOverlay[]) => isControllerExtensionEnabled(overlays, syncViewsIds, isReuseComponentChecker, isCloud)
         });
     } else {
         const extendControllerService = (await import('open/ux/preview/client/adp/extend-controller')).default;