+++ packages/preview-middleware-client/types/sap.ui.fl.d.ts
@@ -82,12 +82,13 @@ declare module 'sap/ui/fl/Scenario' {
 declare module 'sap/ui/fl/Utils' {
     import type ManagedObject from 'sap/ui/base/ManagedObject';
     import type Controller from 'sap/ui/core/mvc/Controller';
-    import type Control from 'sap/ui/core/Control';
+    import type Component from 'sap/ui/core/Component'
 
     interface Utils {
         checkControlId(control: ManagedObject): boolean;
         getViewForControl(control: ManagedObject): ControlView;
-        getAppComponentForControl(control: Control): Control;
+        getAppComponentForControl(control: ManagedObject): Component;
+        getComponentForControl(control: ManagedObject): Component;
     }
 
     interface ControlView {