+++ packages/preview-middleware/src/base/flp.ts
@@ -78,6 +78,9 @@ export class FlpSandbox {
     public readonly test?: TestConfig[];
     public readonly router: EnhancedRouter;
     private fs: MemFsEditor | undefined;
+    private readonly logger: Logger;
+    private readonly utils: MiddlewareUtils;
+    private readonly project: ReaderCollection;
 
     /**
      * Constructor setting defaults and keeping reference to workspace resources.
@@ -87,12 +90,10 @@ export class FlpSandbox {
      * @param utils middleware utilities provided by the UI5 CLI
      * @param logger logger instance
      */
-    constructor(
-        config: Partial<MiddlewareConfig>,
-        private readonly project: ReaderCollection,
-        private readonly utils: MiddlewareUtils,
-        private readonly logger: Logger
-    ) {
+    constructor(config: Partial<MiddlewareConfig>, project: ReaderCollection, utils: MiddlewareUtils, logger: Logger) {
+        this.logger = logger;
+        this.project = project;
+        this.utils = utils;
         this.flpConfig = getFlpConfigWithDefaults(config.flp);
         this.test = config.test;
         this.rta = config.editors?.rta ?? sanitizeRtaConfig(config.rta, logger); //NOSONAR