+++ packages/adp-flp-config-sub-generator/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sap-ux/adp-flp-config-sub-generator
 
+## 0.0.38
+
+### Patch Changes
+
+-   Updated dependencies [fbdf2d0]
+    -   @sap-ux/adp-tooling@0.13.22
+    -   @sap-ux/flp-config-inquirer@0.2.67
+
+## 0.0.37
+
+### Patch Changes
+
+-   Updated dependencies [ced5edf]
+    -   @sap-ux/axios-extension@1.19.2
+    -   @sap-ux/inquirer-common@0.6.32
+    -   @sap-ux/adp-tooling@0.13.21
+    -   @sap-ux/system-access@0.5.34
+    -   @sap-ux/flp-config-inquirer@0.2.66
+
 ## 0.0.36
 
 ### Patch Changes