+++ packages/adp-tooling/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/adp-tooling
 
+## 0.13.22
+
+### Patch Changes
+
+-   fbdf2d0: fix(generator-adp): Generator running in the CLI context
+
+## 0.13.21
+
+### Patch Changes
+
+-   ced5edf: feat(generator-adp): Create a yeoman package for Adaptation Project generator
+-   Updated dependencies [ced5edf]
+    -   @sap-ux/axios-extension@1.19.2
+    -   @sap-ux/inquirer-common@0.6.32
+    -   @sap-ux/system-access@0.5.34
+
 ## 0.13.20
 
 ### Patch Changes