+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/deploy-tooling
 
+## 0.15.51
+
+### Patch Changes
+
+-   Updated dependencies [ced5edf]
+    -   @sap-ux/axios-extension@1.19.2
+    -   @sap-ux/system-access@0.5.34
+
 ## 0.15.50
 
 ### Patch Changes